Rework note spawn

+++ Assets/Scripts/SO/GameInfo.cs
@@ -27,7 +27,7 @@ namespace NsfwMiniJam.SO
         public float MineChancePercent;
         public float HypnotismChance;
         public int HypnotismHitCount;
-        public float HypnotismNextNoteDelay;
+        public int HypnotismNextNoteDelay;
     }
 
     [System.Serializable]
